// Antojitos/ClientApp/src/app/model/sale-interface.ts
export interface SaleInterface {
    id?: string;
    clientId?: string;
    clientName?: string;
    userId?: string;
    userName?: string;
    code?: string;
    iva?: string;
    netoWorth?: string;
    total?: string;
    paymentMethodId?: string;
    paymentName?: string;
    details?: string;
}